# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Trip Service API",
      "description": "API for managing trips, including routes, schedules, and bookings",
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:8083",
        "description": "Development Server"
      },
      {
        "url": "https://api.tripservice.bussewaya.online",
        "description": "Production Server"
      }
    ],
    "paths": {
      "/trip-service/trips": {
        "post": {
          "summary": "Create a new trip",
          "tags": ["Trip"],
          "operationId": "createTrip",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Trip created successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request"
            }
          }
        }
      },
      "/trip-service/trips/{tripId}": {
        "get": {
          "summary": "Get a trip by tripId",
          "tags": ["Trip"],
          "operationId": "getTripById",
          "parameters": [
            {
              "name": "tripId",
              "in": "path",
              "required": true,
              "description": "ID of the trip to fetch",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Trip details",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            },
            "404": {
              "description": "Trip not found"
            }
          }
        },
        "patch": {
          "summary": "Update booking status for a trip",
          "tags": ["Trip"],
          "operationId": "updateBookingStatus",
          "parameters": [
            {
              "name": "tripId",
              "in": "path",
              "required": true,
              "description": "ID of the trip to update",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bookingStatus": {
                      "type": "string",
                      "example": "view"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Booking status updated successfully"
            },
            "404": {
              "description": "Trip not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a trip by tripId",
          "tags": ["Trip"],
          "operationId": "deleteTripById",
          "parameters": [
            {
              "name": "tripId",
              "in": "path",
              "required": true,
              "description": "ID of the trip to delete",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Trip deleted successfully"
            },
            "404": {
              "description": "Trip not found"
            }
          }
        }
      },
      "/trip-service/trips/{startLocation}/{endLocation}/{tripDate}": {
        "get": {
          "summary": "Get all trips by location and date",
          "tags": ["Trip"],
          "operationId": "getTripsByLocationAndDate",
          "parameters": [
            {
              "name": "startLocation",
              "in": "path",
              "required": true,
              "description": "Starting location of the trip",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "endLocation",
              "in": "path",
              "required": true,
              "description": "Ending location of the trip",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "tripDate",
              "in": "path",
              "required": true,
              "description": "The date of the trip",
              "schema": {
                "type": "string",
                "format": "date"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of trips",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Trip"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameters"
            }
          }
        }
      },
      "/trip-service/trips/{scheduleId}/{tripDate}": {
        "get": {
          "summary": "Get all trips by scheduleId and tripDate",
          "tags": ["Trip"],
          "operationId": "getTripsByScheduleAndDate",
          "parameters": [
            {
              "name": "scheduleId",
              "in": "path",
              "required": true,
              "description": "Schedule ID to retrieve trips for",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "tripDate",
              "in": "path",
              "required": true,
              "description": "The date of the trip",
              "schema": {
                "type": "string",
                "format": "date"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of trips",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Trip"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid query parameters"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Trip": {
          "type": "object",
          "properties": {
            "tripId": {
              "type": "number"
            },
            "tripNumber": {
              "type": "string"
            },
            "tripDate": {
              "type": "string",
              "format": "date-time"
            },
            "bookingStatus": {
              "type": "string"
            },
            "confirmedSeatsCount": {
              "type": "number"
            },
            "availableSeatsCount": {
              "type": "number"
            },
            "routeNumber": {
              "type": "string"
            },
            "startLocation": {
              "type": "string"
            },
            "endLocation": {
              "type": "string"
            },
            "scheduleId": {
              "type": "number"
            },
            "departureTime": {
              "type": "string"
            },
            "arrivalTime": {
              "type": "string"
            },
            "permitNumber": {
              "type": "string"
            },
            "vehicleNumber": {
              "type": "string"
            },
            "busType": {
              "type": "string"
            },
            "pricePerSeat": {
              "type": "number"
            },
            "music": {
              "type": "boolean"
            },
            "ac": {
              "type": "boolean"
            }
          },
          "required": ["tripId", "tripNumber", "tripDate", "bookingStatus", "confirmedSeatsCount", "availableSeatsCount", "routeNumber", "startLocation", "endLocation", "scheduleId", "permitNumber"]
        }
      }
    }
  }
  
  